Refactoring storage modules
