+++ wallet/src/Ergvein/Wallet/Storage/Data.hs
@@ -1,13 +1,16 @@
-module Ergvein.Wallet.Storage.Secure.Data
+module Ergvein.Wallet.Storage.Data
   (
     WalletData(..)
   , EncryptedWalletData(..)
+  , ErgveinStorage(..)
   ) where
 
 import Ergvein.Aeson
 import Ergvein.Crypto
 import Data.Sequence
 import Data.Text
+
+import qualified Data.Map.Strict as M
 
 data WalletData = WalletData
   { mnemonic    :: Mnemonic
@@ -21,3 +24,14 @@ data EncryptedWalletData = EncryptedWalletData
   , initVector      :: Text
   } deriving (Show)
 $(deriveJSON defaultOptions ''EncryptedWalletData)
+
+data ErgveinStorage = ErgveinStorage{
+  storage'wallet     :: EncryptedWalletData
+, storage'pubKeys    :: M.Map EgvXPubKey [Base58]
+, storage'walletName :: Text
+}
+
+instance Eq ErgveinStorage where
+  a == b = storage'walletName a == storage'walletName b
+
+$(deriveJSON aesonOptionsStripToApostroph ''ErgveinStorage)
