+++ wallet/src/Ergvein/Wallet/Storage/Util.hs
@@ -1,4 +1,4 @@
-module Ergvein.Wallet.Storage.Secure(
+module Ergvein.Wallet.Storage.Util(
     WalletData(..)
   , EncryptedWalletData(..)
   , createWallet
@@ -9,8 +9,10 @@ module Ergvein.Wallet.Storage.Secure(
   , addXPrvKey
   , getWalletDirectory
   , walletFileName
+  , loadStorageFromFile
   ) where
 
+import Control.Monad.IO.Class
 import Data.ByteArray           (convert)
 import Data.ByteString          (ByteString)
 import Data.ByteString.Base64   (encode, decodeLenient)
@@ -21,10 +23,17 @@ import Data.Text.Encoding.Error
 import Ergvein.Aeson
 import Ergvein.Crypto
 import Ergvein.IO
-import Ergvein.Wallet.Input     (Password)
-import Ergvein.Wallet.Storage.Secure.Data
+import Ergvein.Text
+import Ergvein.Wallet.Language
+import Ergvein.Wallet.Localization.Native
+import Ergvein.Wallet.Native
+import Ergvein.Wallet.Storage.Data
 import System.Directory
 import System.FilePath
+
+import qualified Data.Text as T
+
+type Password = Text
 
 -- | Get user home directory location with ".ergvein" directory appended.
 getWalletDirectory :: IO FilePath
@@ -94,3 +103,31 @@ decryptWalletData encryptedWalletData password =
     secretKey = Key (fastPBKDF2_SHA256 defaultPBKDF2Params (encodeUtf8 password) saltBS) :: Key AES256 ByteString
     initIV = makeIV $ decodeLenient $ encodeUtf8 $ initVector encryptedWalletData :: Maybe (IV AES256)
     encryptedDataBS = decodeLenient $ encodeUtf8 $ encryptedData encryptedWalletData
+
+
+storageFileName :: Text
+storageFileName = "storage"
+
+-- TODO: Actually decrypt the storage
+decryptStorage :: MonadIO m => Password -> Text -> m (Either StorageAlerts ErgveinStorage)
+decryptStorage pass txt = pure $ either (Left . SADecodeError) Right $ text2json txt
+
+loadStorageFromFile :: (MonadIO m, HasStoreDir m, PlatformNatives) => Password -> m (Either StorageAlerts ErgveinStorage)
+loadStorageFromFile pass = do
+  storageResp <- readStoredFile storageFileName
+  either (pure . Left . SANativeAlert) (decryptStorage pass . T.concat) storageResp
+
+-- Alerts regarding secure storage system
+data StorageAlerts
+  = SADecodeError Text
+  | SALoadedSucc
+  | SANativeAlert NativeAlerts
+  deriving (Eq)
+
+instance LocalizedPrint StorageAlerts where
+  localizedShow l v = case l of
+    English -> case v of
+      SADecodeError e -> "Storage loading error: " <> e
+      SALoadedSucc    -> "Storage loaded"
+      SANativeAlert a -> localizedShow l a
+    Russian -> localizedShow English v
